Move `vctrs_split_id()` to `split.c`
